requirements.txt
discord.py>=0.16.12
twython>=3.6.0